Moved Simple JSON stuff into admin.json, finished up the Admin side JSON.
Whomever does the BB side, you can literally just copy mode of the code from the admin side. Just have to use more primitives and J2ME friendly stuff.

As far as data saving etc, all that needs to be done is read and write to the file now. All strings are created using the toJSONString() and on any data.* object call fromJSONString() and it will parse into the object.
This operation CLOBBERS ALL DATA.