Added some bug fixes.
Fixed contestant IDs to be two chars (ReGeX and input)
Fixed row select to that it now updates the ID.

Added exceptions to be thrown when setting fields. This allows us to
know when we set something incorrectly, it'll be thrown. Forces us to
add more try..catch blocks, but it adds better functionality.

Updated some data types I think..